Add functions to Piece to return moves
The functions moves and capture_moves will return a vector consisting of
vectors of directions. For most pieces the inner vector will only contain
a single element. For pieces like the knight it will contain two directions.

The capture_moves function is often also the same as the moves function,
since most pieces capture using the same move as their normal move,
with the exception of the pawn.