Update Board to store Squares instead of Pieces
This will make the code a bit more logical.
Square now contains a function to check if it is empty.
The board also has two new functions, get() and get_mut(),
which will return a (mutable) reference to a Square matching the rank
and file of the Square supplied as argument.